Fix bugs in compare functions in example
test_rb_compare and test_avl_compare were casting directly
from node* to the container*.  That's wrong.  Need to subtract
the offset of the node member which is what the provided macros
are for.

All the macros are actually identical.  Could just have one and
call it container_of() like the linux kernel.