Used clang-tidy to add braces around all statements.
Clang-tidy (http://clang.llvm.org/extra/clang-tidy/) can automatically modify
C code so that it complies with certain best practices.  In this case, I used
it to put braces around all single-line statements, such as `if` statements
that guard a single statement.

The reason for this is that a common error is to accidentally assume that
you've got a statement inside of a block guarded by an `if` statement, only to
discover that you forgot the braces, leaving you with nothing.  This change
ensures that it is **very** obvious as to which statements are in the block,
and which aren't.