Partial bug fix: `list_push_front()` and `list_push_back()` now remove nodes before pushing.
The original code pushed the link into the list, without removing it from the list.  As a result, if you pushed the same link into a list twice, the link's original neighbors would end up pointing to the wrong thing (the moved link instead of each other).

The reason that this is only a partial bug fix is because if you are moving a link from one list to a different list, you must remove the link from it's original list first.  However, the current code cannot support that; it assumes that the link is a member of the list it's being removed from.  There are two possible fixes for this; first, we can add in an extra parameter to every function so we know which list it's being moved from.  This will break all code that depends on this not happening.  Second, we can add another parameter to the `link_s` structure that points to the list the link is currently a part of.  This will break less code, and so is probably the better choice.