Partial bug fix: links are now removed from lists before being inserted.
In the original code, links were simply inserted into lists by `list_insert_before()` and `list_insert_after()`.  This only works if the link is not already a part of some list.  If it is, then the link's neighbors will be confused and end up pointing at a link that is no longer their neighbor.  This commit fixes that bug by removing the node first (which fixes the link's neighbors), and then adding the link into where it belongs.

Unfortunately, this commit still suffers from the previous commit; if you move a link from one list to a different list, the remove function doesn't work correctly.  The least disruptive move is to add a pointer to the link that points to the list it is currently a member of.  See the previous commit message for more details.